# schema/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FAR-Lab API",
    "description": "FAR-Lab 对外 HTTP API（24§5）。V2 receipts 端点采用统一信封 { ok: true, data: T } + RFC 7807 错误响应（R-05）。",
    "version": "2026-06-27"
  },
...
  "paths": {
...
    "/api/v1/planning/risk": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "readOnly": {
                    "type": "boolean"
                  },
                  "docOnly": {
                    "type": "boolean"
                  },
                  "boundedWrite": {
                    "type": "boolean"
                  },
                  "touchesTrustKernel": {
                    "type": "boolean"
                  },
                  "newCliOrApi": {
                    "type": "boolean"
                  },
                  "crossModule": {
                    "type": "boolean"
                  },
                  "destructive": {
                    "type": "boolean"
                  },
                  "irreversible": {
                    "type": "boolean"
                  },
                  "ambiguous": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "readOnly",
                  "docOnly",
                  "boundedWrite",
                  "touchesTrustKernel",
                  "newCliOrApi",
                  "crossModule",
                  "destructive",
                  "irreversible",
                  "ambiguous"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "level": {
                          "type": "string",
                          "enum": [
                            "P0",
                            "P1",
                            "P2",
                            "P3",
                            "P4"
                          ]
                        },
                        "reasons": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "level",
                        "reasons"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "source_anchor": {
                      "type": "object",
                      "properties": {
                        "fileId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "stageId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "callRecordId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      },
                      "required": [
                        "fileId",
                        "stageId",
                        "callRecordId"
                      ],
                      "additionalProperties": false
                    },
                    "detail": {}
                  },
                  "required": [
                    "error_code",
                    "message",
                    "source_anchor"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planning/plan": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "goal": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "steps": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "action": {
                          "type": "string",
                          "minLength": 1
                        },
                        "risk": {
                          "type": "string",
                          "enum": [
                            "P0",
                            "P1",
                            "P2",
                            "P3",
                            "P4"
                          ]
                        },
                        "tools": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "minLength": 1
                          },
                          "default": [
                            "Bash"
                          ]
                        },
                        "dependsOn": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "minLength": 1
                          },
                          "default": []
                        },
                        "verification": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "id",
                        "action",
                        "risk",
                        "verification"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1
                  },
                  "taskClass": {
                    "type": "string"
                  },
                  "lifecycleDepth": {
                    "type": "string"
                  }
                },
                "required": [
                  "goal",
                  "steps"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "ok": {
                          "type": "boolean"
                        },
                        "violations": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "stepId": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "stepId",
                              "code",
                              "message"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "executionOrder": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "ok",
                        "violations",
                        "executionOrder"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "source_anchor": {
                      "type": "object",
                      "properties": {
                        "fileId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "stageId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "callRecordId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      },
                      "required": [
                        "fileId",
                        "stageId",
                        "callRecordId"
                      ],
                      "additionalProperties": false
                    },
                    "detail": {}
                  },
                  "required": [
                    "error_code",
                    "message",
                    "source_anchor"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planning/spec": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "story": {
                    "type": "string",
                    "minLength": 1
                  },
                  "delta": {
                    "type": "object",
                    "properties": {
                      "added": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "minLength": 1
                        },
                        "default": []
                      },
                      "modified": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "minLength": 1
                        },
                        "default": []
                      },
                      "removed": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "minLength": 1
                        },
                        "default": []
                      }
                    },
                    "additionalProperties": false
                  },
                  "acceptanceCriteria": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "statement": {
                          "type": "string",
                          "minLength": 1
                        },
                        "verification": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "id",
                        "statement",
                        "verification"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1
                  },
                  "dataFlow": {
                    "type": "string"
                  },
                  "risk": {
                    "type": "string",
                    "enum": [
                      "P0",
                      "P1",
                      "P2",
                      "P3",
                      "P4"
                    ]
                  },
                  "trustKernel": {
                    "type": "object",
                    "properties": {
                      "additiveOnly": {
                        "type": "boolean"
                      },
                      "cannotProveStatement": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "additiveOnly",
                      "cannotProveStatement"
                    ],
                    "additionalProperties": false
                  },
                  "removedJustification": {
                    "type": "string"
                  },
                  "taskId": {
                    "type": "string"
                  }
                },
                "required": [
                  "story",
                  "delta",
                  "acceptanceCriteria",
                  "risk"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "ok": {
                          "type": "boolean"
                        },
                        "violations": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "message"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "ok",
                        "violations"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "source_anchor": {
                      "type": "object",
                      "properties": {
                        "fileId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "stageId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "callRecordId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      },
                      "required": [
                        "fileId",
                        "stageId",
                        "callRecordId"
                      ],
                      "additionalProperties": false
                    },
                    "detail": {}
                  },
                  "required": [
                    "error_code",
                    "message",
                    "source_anchor"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planning/gate": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "command": {
                          "type": "string",
                          "minLength": 1
                        },
                        "expected": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "command",
                        "expected"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "results": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "actual": {
                          "type": "string",
                          "minLength": 1
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pass",
                            "fail",
                            "not_run"
                          ]
                        }
                      },
                      "required": [
                        "actual",
                        "status"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "items",
                  "results"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "command": {
                                "type": "string"
                              },
                              "expected": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "command",
                              "expected"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "results": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "object",
                            "properties": {
                              "actual": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string",
                                "enum": [
                                  "pass",
                                  "fail",
                                  "not_run"
                                ]
                              }
                            },
                            "required": [
                              "actual",
                              "status"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "passed": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "notRun": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "conclusion": {
                          "type": "string",
                          "enum": [
                            "DONE",
                            "IMPLEMENTED_UNVERIFIED",
                            "BLOCKED"
                          ]
                        },
                        "rationale": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "items",
                        "results",
                        "passed",
                        "failed",
                        "notRun",
                        "conclusion",
                        "rationale"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "source_anchor": {
                      "type": "object",
                      "properties": {
                        "fileId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "stageId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "callRecordId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      },
                      "required": [
                        "fileId",
                        "stageId",
                        "callRecordId"
                      ],
                      "additionalProperties": false
                    },
                    "detail": {}
                  },
                  "required": [
                    "error_code",
                    "message",
                    "source_anchor"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}